Changed Gambia,The and Bahamas,The to "The .."
The ',The' convention crept into the name field as we prepared for launch,
This is not a convention we want to establish for names of companies, schools,
charities, etc. Consumers who need to sort county names ignoring definite or
indefinite articles will should transform the data before performing their
collation, which may have domain specific rules.